+++ changelog.md
@@ -6,6 +6,7 @@
 This version is not released yet and is under active development.
 ```
 
+- \[brew,cask\] Do not let homebrew auto-update on other commands. Refs {issue}`36`.
 - \[brew,cask\] Disable analytics and env hints in logs.
 - \[xbar\] Fix log verbosity and unittests for xbar plugin.
 - \[mpm\] Show in debug logs the extra environment variable used for CLIs.